luv-netboot-image: use a single '=' when comparing strings
When deciding how to create the symbolic link for the network-bootable
image, we compare the TARGET_ARCH variable vs aarch64. However, using
"==" only works in bash. bitbake uses a more generic sh syntax. This
causes the conditional to always fall in the 'else' case and, therefore,
not creating the link correctly.

Cc: Naresh Bhat <naresh.bhat@linaro.org>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>