linux-yocto-efi-test: Automate retrieval of Linux defconfig
1.Define KBUILD_DEFCONFIG variable
The variable KBUILD_DEFCONFIG defined as defconfig refer the link for more details
https://github.com/01org/luv-yocto/blob/master/meta/classes/kernel-yocto.bbclass#L102

2. Define KCONFIG_MODE as alldefconfig
This will select the default option of kernel configurations. Otherwise (i.e.,
using allnoconfig), we lose all those configs that we don't explicitly select.
The arch-specific default configuration files are not exhaustive. When
executing make <arch>_defconfig. It seems that make fills the options
not specified in the config file with the default value. You can refer merge_config script
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/scripts/kconfig/merge_config.sh

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Naresh Bhat <naresh.bhat@linaro.org>