linux-firmware: Get rid of License warnings
When we do a clean build, we see a lot of warnings related to missing
licenses as follows:
WARNING: LICENSE_linux-firmware-ralink-license includes licenses
(Firmware-ralink-firmware) that are not listed in LICENSE
WARNING: LICENSE_linux-firmware-ralink includes licenses
(Firmware-ralink-firmware) that are not listed in LICENSE
WARNING: LICENSE_linux-firmware-radeon-license includes licenses
(Firmware-radeon) that are not listed in LICENSE
WARNING: LICENSE_linux-firmware-radeon includes licenses
(Firmware-radeon) that are not listed in LICENSE

This is because we are overwriting the "LICENSE" variable but not the
"PACKAGES" variable in the .bbappend file in the meta-layer. As such,
bitbake does not find the corresponding LICENSES listed for the various
packages.

Currently, we install a lot of firmware related files and remove those
that are not needed by any test-suite LUV uses. Instead, we create and
install only the new package linux-firmware-luv, which contains only
the license files that LUV makes use off.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>