systemd: Use ConditionKernelCommandLine to enable tty1
Presently, we use a service file splash_tty.service to run a script
splash_runner which checks if "splash" is passed on the command line
and accordingly creates a temporary directory.
The getty@.service.m4 file conditionally executes only if this directory
is present using the ConditionPathIsDirectory construct. This seems
convoluted and can be replaced by simply using the
ConditionKernelCommandLine construct instead.

Signed-off-by: Megha Dey <megha.dey@linux.intel.com>