luv-tests: Configure all the testsuites in LUV
With "noluv" parameter, LUV disables not one but all testsuites which is
not adequate when user wants to disable specific testsuite(s).

Configure all the testsuites such that each testsuite is enabled or
disabled independent of other testsuites.

Ideally, all the testsuites are enabled by default. Disable a testsuite
or two if needed by just removing the string that represents the
testsuite in "LUV_TESTS" parameter in luv.cfg file.

Cc: Ricardo Neri <ricardo.neri@intel.com>
Signed-off-by: Gayatri Kammela <gayatri.kammela@intel.com>